% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc.R
\name{alloc}
\alias{alloc}
\title{Allocate Required Storage}
\usage{
alloc(x, data, gmax, par.env = parent.frame())
}
\arguments{
\item{x}{the result of preprocessing an expression}
}
\description{
Allocate Required Storage
}
