% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc.R
\name{alloc_temp}
\alias{alloc_temp}
\title{Track Required Allocations for Intermediate vectors}
\usage{
alloc_temp(temp, depth, size)
}
\description{
List of all allocated temporary vector sizes, and which of those are free.
}
