% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoexec.R
\name{autoexec}
\alias{autoexec}
\title{Generate an Iterative Self Executing \code{r2c} Function}
\usage{
autoexec(fun, runner, ...)
}
\arguments{
\item{fun}{an \code{r2c} function.}

\item{runner}{an \code{r2c} runner function.}

\item{...}{additional parameters to pre-set for the runner.}

\item{enclos}{an environment to use as the parent frame for the \code{r2c}
function evaluation.}
}
\description{
Normally \code{{r2c}} functions are invoked via runners functions such as
\code{\link{group_exec}}.  For the case where a function is frequently executed with
the same runner, \code{autoexec} can generate a new function that combines the
runner and the \code{{r2c}} function into one.
}
