% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{base_grp_eval}
\alias{base_grp_eval}
\title{Basic Split Apply Combine}
\usage{
base_grp_eval(data, g, call, env = parent.frame())
}
\arguments{
\item{data}{a data frame with numeric columns}

\item{call}{quoted R call to evaluate}

\item{env}{environment to use as enclosure}
}
\description{
Evaluates quoted expressions in the context of data split by group.  Intended
purely for testing against C calculations.
}
