% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bsac}
\alias{bsac}
\title{Basic Split Apply Combine}
\usage{
bsac(call, g, data, MoreArgs, env = parent.frame())
}
\arguments{
\item{call}{quoted R call to evaluate}

\item{data}{a data frame with numeric columns}

\item{env}{environment to use as enclosure}
}
\value{
numeric vector
}
\description{
Evaluates quoted expressions in the context of data split by group.  Intended
purely for testing against C calculations.
}
