% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{cmp}
\alias{cmp}
\alias{cmpq}
\title{Compile an R Call Into Machine Instructions}
\usage{
cmp(call, dir = tempfile())

cmpq(call, dir = tempfile())
}
\arguments{
\item{call}{an R expression, for \code{compileq} it is captured unevaluated, for
\code{compile} it should be pre-quoted.}

\item{dir}{character(1L) name of a temporary directory to store the shared
object file in.}
}
\value{
the full path to the compiled shared object file.
}
\description{
Translates an R call into C and compiles it into native instructions using
\verb{R CMD SHLIB}.
}
\examples{
compileq(sum(x + y))
compile(quote(sum(x + y))
}
