% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{code_gen}
\alias{code_gen}
\title{Code Generation Functions}
\usage{
code_gen(dat)
}
\description{
\itemize{
\item Must accept one of four possible signatures.
\item Must set result by reference.
\item Must set result size by reference.
}
}
