% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc.R
\name{fap_fun}
\alias{fap_fun}
\title{Initializer for Function Registration Entries}
\usage{
fap_fun(name, fun, defn, ctrl.params = character(), type)
}
\arguments{
\item{name}{character(1L) symbol that will reference the function}

\item{fun}{the function we're trying to emulate}

\item{defn}{NULL if fun is a closure, otherwise a function template to use
for \code{\link{match.call}}'s \code{definition} paramter.}

\item{ctrl.params}{character names of all the formal parameters that are
to be evaluated once up front and not for each group in the data.  If any
data columns are matched to these parameters, the entire data column will
be used for them.}

\item{type}{list(2L) containing the type of function in "constant", "arglen",
or "vecrec" at position one, and additional meta data at position two
described next:
\itemize{
\item constant: a positive non-NA integer indicating the constant result size
(e.g. 1L for \code{mean})
\item arglen: character(1L) the name of the argument to use the length of as
the result size (e.g. \code{probs} for \code{\link{quantile}}.
\item vecrec: character(n) the names of the arguments to use to compute result
size under assumption of recycling to longest, or zero if any argument is
zero length.
}}
}
\value{
a list containing the above information after validating it.
}
\description{
Initializer for Function Registration Entries
}
