% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{first_vec}
\alias{first_vec}
\title{Retrieve First Vector from Data}
\usage{
first_vec(x)
}
\arguments{
\item{x}{a numeric vector, or a (possibly empty) list of numeric vectors.}
}
\value{
if \code{data} is a list, the first element if it is a numeric vector, or
an empty numeric vector if the list is empty.  If \code{data} is a numeric
vector, then \code{data}.  Otherwise an error is thrown.
}
\description{
Designed to handle the case where \code{data} can be either a numeric vector or a
list of numeric vectors.
}
\examples{
first_vec(1:5)
first_vec(runif(5))
first_vec(mtcars)
first_vec(matrix(1:4, 2))  # matrices treated as vectors
}
