% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{get_c_code}
\alias{get_c_code}
\alias{get_r_code}
\alias{get_so_loc}
\alias{get_compile_out}
\title{Extract Data from "r2c_fun" Objects}
\usage{
get_c_code(r2c.fun)

get_r_code(r2c.fun)

get_so_loc(r2c.fun)

get_compile_out(r2c.fun)
}
\arguments{
\item{r2c.fun}{an "r2c_fun" function as generated by e.g. \code{\link{r2c}}.}
}
\value{
for \code{get_r_code} the R language object the C code is based on,
otherwise a character vector.
}
\description{
"r2c_fun" functions contain embedded data used by the runners to call the
compiled native code associated with the functions.  The \verb{get_*} functions
documented here extract various aspects of this data.
}
\details{
\itemize{
\item \code{get_so_loc} the file system location of the shared object file.
\item \code{get_c_code} the generated C code used to produced the shared object.
\item \code{get_r_code} the R call that was translated into the C code.
\item \code{get_compile_out} the "stdout" produced during the compilation of the
shared object.
}
}
\examples{
r2c_sum_sub <- r2cq(sum(x + y))
get_r_code(r2c_sum_sub)
writeLines(get_c_code(r2c_sum_sub))
}
\seealso{
\code{\link{r2c}}.
}
