% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{group_exec}
\alias{group_exec}
\title{Execute r2c Function Iteratively on Groups in Data}
\usage{
group_exec(fun, groups, data, MoreArgs = list(), sorted = FALSE)
}
\arguments{
\item{fun}{a function produced by \code{r2c}.}

\item{groups}{an integer vector, or a list of equal-length integer vectors,
the interaction of which defines individual groups to organize the data
into.  Support for non-integer vectors will be added in the future.}

\item{data}{a numeric vector, or a list of numeric vectors, each vector the
same length as the vector(s) in \code{groups}.  If a named list, the vectors
will be matched to \code{fun} parameters by those names.  Elements without names
are matched positionally.}

\item{MoreArgs}{a list of R objects to pass on as non-data parameters to
\code{fun}.  Unlike with \code{data}, the entire object will be passed for each
group.  This is useful for arguments that are intended to remain constant
group to group.}

\item{sorted}{TRUE or FALSE (default), whether the vectors in \code{data} and
\code{groups} are already sorted by \code{groups}.  If set to TRUE, the \code{data} will
not be sorted prior to computation; if the data is truly sorted this
produces the same results while avoiding the cost of sorting.}
}
\value{
a if \code{groups} and \code{data} are atomic vectors, a named numeric vector
with the results of executing \code{fun} on each group and the names set to the
groups.  Otherwise, a "data.frame" with the group vectors as columns and
the result column last.
}
\description{
Organizes \code{data} into row groups according to \code{groups}, and calls \code{r2c}
functions for each group with the arguments bound to the portion of the data
corresponding to each group.
}
