% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-loop.R, R/code-ifelse.R,
%   R/code-assign-braces.R
\name{r2c_for}
\alias{r2c_for}
\alias{for_init}
\alias{for_iter}
\alias{for_n}
\alias{for_0}
\alias{lset}
\alias{lrec}
\alias{r2c_if}
\alias{if_true}
\alias{if_false}
\alias{if_test}
\alias{vcopy}
\alias{rec}
\title{Intermediate Representation Functions}
\usage{
r2c_for(iter, for.n, for.0)

for_init(seq, seq.i)

for_iter(var, seq, seq.i)

for_n(expr)

for_0(expr)

lset(rec.i, x)

lrec(rec.i, x)

r2c_if(true, false)

if_true(expr)

if_false(expr)

if_test(cond)

vcopy(x)

rec(x)
}
\arguments{
\item{iter}{expression that increments the loop iteration variable.}

\item{for.n}{expression to evaluate if there is at least one loop iteration.}

\item{for.0}{expression to evaluate if there are no loop iterations.}

\item{seq}{expression that generates vector to loop over.}

\item{seq.i}{position in \code{seq} the loop is at.}

\item{var}{variable to hold the loop iteration value.}

\item{expr}{branch expression to evaluate.}

\item{rec.i}{identifier for each instance of a use before set loop variable.}

\item{x}{language to wrap in markers that designate it for special treatment
by the memory allocator at run time.}

\item{true}{expression to evaluate if previous \code{if_test} is TRUE.}

\item{false}{expression to evaluate if previous \code{if_test} is FALSE}

\item{cond}{expression to evaluate to determine which branch is taken.}
}
\value{
These functions are not intended for use at the R level thus their
return values are undocumented.
}
\description{
Functions used by the \code{r2c} preprocessor to represent control structures in a
format that facilitates translation to C code.
}
\seealso{
\link[=r2c-preprocess]{Preprocessing}
}
\keyword{internal}
