% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{make_shlib}
\alias{make_shlib}
\title{Generate a Shared Library Object From C Code}
\usage{
make_shlib(x)
}
\arguments{
\item{x}{character the C code to compile}
}
\value{
character file name of the SO; will be in a temporary directory, it is
the users responsibility to preserve and/or discard the file.
}
\description{
Generate a Shared Library Object From C Code
}
