% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-summary.R
\name{mean1}
\alias{mean1}
\title{Single Pass Mean Calculation}
\usage{
mean1(x, na.rm = FALSE)
}
\arguments{
\item{x}{An \R object.  Currently there are methods for
    numeric/logical vectors and \link[base:Dates]{date},
    \link[base]{date-time} and \link[base]{time interval} objects.  Complex vectors
    are allowed for \code{trim = 0}, only.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE}
    indicating whether \code{NA} values should be stripped before the
    computation proceeds.}
}
\value{
scalar numeric
}
\description{
\code{\link[base:mean]{base::mean}} does a two pass calculation that additional handles cases
where \code{sum(x)} overflows doubles but \code{sum(x/n)} does not.  This version is a
single pass one that does not protect against the overflow case.
}
\examples{
mean1(runif(10))

## Overflow even 80 bit long double:
mean1(rep(.Machine$double.xmax, 2^4))
mean(rep(.Machine$double.xmax, 2^4))

## Reduced precision
x <- runif(10) ^ 2
mean(x) - mean1(x)
}
