% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-numeric.R
\name{numeric_along}
\alias{numeric_along}
\alias{numeric_alongn}
\title{Initialize a Numeric Vector Sized to Match Input}
\usage{
numeric_along(along.with)

numeric_alongn(...)
}
\arguments{
\item{along.with}{vector to use for sizing the result.}

\item{...}{vectors to use for sizing the result.}
}
\value{
a zero numeric vector the same length as the product of the lengths
of all the vectors in provided as inputs.
}
\description{
Generates a numeric vector of the same size as the input.  Equivalent to
\code{numeric(length(x))}.  \code{numeric_alongn} supports multiple vectors in the
input, for which the result size is the product of the lengths of the inputs.
}
\examples{
numeric_along(1:3)
numeric_alongn(1:3, 1:2)
}
\seealso{
\link[base:numeric]{base::numeric}
}
