% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc.R
\name{prepare}
\alias{prepare}
\title{Prepare Calls for Evaluation}
\usage{
prepare(call, data, g, env, depth, alloc.dat, code)
}
\arguments{
\item{call}{an unevaluated R call}

\item{data}{character the names of the data columns}

\item{g}{max group size}

\item{env}{calling frame to evaluate expression in}

\item{depth}{how deep into the expression we've recursed}

\item{alloc.dat}{structure that tracks the required size}

\item{code}{a list of}
}
\value{
a list containing (update):
\itemize{
\item Accrued temporary storage requirements.
\item Size of the current call being assessed.
\item A linear list of generated code and associated calls.
}
}
\description{
Retrieve appropriate entry points based on configuration parameters, and
compute required temporary storage sizes.
}
