% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Generate C Code for Compilation}
\usage{
preprocess(call)
}
\arguments{
\item{call}{an unevaluated R call}
}
\value{
a list containing (update):
\itemize{
\item Accrued temporary storage requirements.
\item Size of the current call being assessed.
\item A linear list of generated code and associated calls.
}
}
\description{
Match each R symbol to C code, possibly parameterized by some simple
parameters such as which parameters are specified, but not much else as
we need to know the groupwise data to make more decisions.
}
