% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{r2c-inspect}
\alias{r2c-inspect}
\alias{get_c_code}
\alias{get_r_code}
\alias{get_so_loc}
\alias{get_compile_out}
\alias{show_c_code}
\title{Extract Data from "r2c_fun" Objects}
\usage{
get_c_code(fun, all = TRUE)

get_r_code(fun, raw = FALSE)

get_so_loc(fun)

get_compile_out(fun)

show_c_code(fun, all = FALSE)
}
\arguments{
\item{fun}{an "r2c_fun" function as generated by e.g. \code{\link{r2c}}.}

\item{all}{TRUE or FALSE (default) whether to retrieve all of the C code, or
just the portion directly corresponding to the translated R expression.}

\item{raw}{TRUE or FALSE (default) whether to display the processed R code
exactly as \code{r2c} will use it, or to simplify it to make easier to read.  If
a simplification occurred the processed member name will be "processed*".}
}
\value{
for \code{get_r_code} a list with on or two members, the first "original"
is the R language object provided to the \link[=r2c-compile]{compilation functions}, the second "processed" (or "processed*", see \code{raw}
parameter) is the version that the C code is based on.  For all other
functions a character vector, invisibly for \code{show_c_code}.
}
\description{
"r2c_fun" functions contain embedded data used by the runners to call the
compiled native code associated with the functions.  The functions documented
here extract various aspects of this data.
}
\details{
\itemize{
\item \code{get_so_loc} the file system location of the shared object file.
\item \code{get_c_code} the generated C code used to produce the shared object, but
for quick inspection \code{show_c_code} is best.
\item \code{show_c_code} retrieves code with \code{get_c_code} and outputs to screen the
portion corresponding to the compiled expression, or optionally all of it.
\item \code{get_r_code} the R call that was translated into the C code; if
processing modified the original call the processed version will also be
shown (see \code{\link{r2cq}}).
\item \code{get_compile_out} the "stdout" produced during the compilation of the
shared object.
}

Most calls seen in the raw version of what \code{get_r_code} returns will have a C
level counterpart labeled with the R call in a comment.  This includes calls
that are nested as arguments to other calls, which will appear before the
outer call.  Due to how how control structures are implemented the R calls
and the C level counterparts will not match up exactly.
}
\examples{
r2c_sum_sub <- r2cq(sum(x + y))
get_r_code(r2c_sum_sub)
show_c_code(r2c_sum_sub)
}
\seealso{
\code{\link{r2c-compile}}, \code{\link{r2c-preprocess}}.
}
