% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2c-package.R
\docType{package}
\name{r2c-package}
\alias{r2c-package}
\alias{r2c}
\title{Fast Iterated Statistic Computation in R}
\description{
Compiles a subset of R into machine code so that expressions composed with
that subset can be applied repeatedly on varying data without interpreter
overhead.  See the \link{group_exec} examples for a quick start.
}
\details{
Basics:
\itemize{
\item \link[=r2c-supported-funs]{Supported functions}: which R functions \code{r2c} can
compile.
\item \link[=r2c-compile]{Compilation facilities}: how to compile R with \code{r2c}.
\item \link[=runners]{Runners}, to execute your code iteratively:
\itemize{
\item By \link[=group_exec]{group}.
\item Across \link[=rolli_exec]{windows}.
}
}

Details:
\itemize{
\item \link[=r2c-performance]{Performance considerations}.
\item \link[=r2c-memory]{Memory use}.
\item \link[=r2c-preprocess]{Preprocessing}.
\item \link[=r2c-control-structures]{Control structures} (experimental).
\item \link{Expression Types}: to understand where it is allowable to use
non-supported functions and why some supported function parameters are
required to be constant.
}
}
