% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-assign-control.R
\name{r2c_copy}
\alias{r2c_copy}
\title{Copy a Vector Into a New One}
\usage{
r2c_copy(x)
}
\arguments{
\item{x}{numeric vector to copy}
}
\value{
\code{x}, copied to a new location
}
\description{
Copies a vector into a new memory location.  This function is intended solely
for documentation purposes of calls generated by \code{r2c} as part of the
compilation process.  In some cases, it is necessary for \code{r2c} to explicitly
copy a variable into the result vector.  This is achieved by the C
counterpart of this function.  Since all other \code{r2c} C functions have R
counterparts, we add this R counterpart for consistency.  There is no reason
for you to ever use this function.
}
