% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-loop.R
\name{r2c_for}
\alias{r2c_for}
\alias{for_init}
\alias{for_iter}
\alias{for_n}
\alias{for_0}
\title{Loop Stub Functions}
\usage{
r2c_for(iter, for.n, for.0)

for_init(seq, seq.i)

for_iter(var, seq, seq.i)

for_n(expr.n)

for_0(expr.0)
}
\description{
Loop Stub Functions
}
\keyword{internal}
