% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{r2c_lib}
\alias{r2c_lib}
\title{Create a Library of Pre-Compiled "r2c_fun" Objects}
\usage{
r2c_lib(
  xs,
  check = getOption("r2c.check.result", FALSE),
  quiet = getOption("r2c.quiet", TRUE),
  optimize = getOption("r2c.optimize", TRUE),
  envir = lapply(xs, environment)
)
}
\arguments{
\item{xs}{named list of R functions to convert to "r2c_fun" objects.}
}
\value{
and "r2c_lib" object.
}
\description{
Create a Library of Pre-Compiled "r2c_fun" Objects
}
\examples{
lib.file <- tempfile()
saveRDS(
  r2c_lib(list(sum=function(x) sum(x), mean=function(x) mean(x))),
  lib.file
)
lib <- readRDS(lib.file)
lib$sum(1:10)
lib$mean(1:10)
with(mtcars, group_exec(lib$mean, hp, cyl))
}
