% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-assign-control.R
\name{rec}
\alias{rec}
\title{Mark a Vector for Reconciliation}
\usage{
rec(x)
}
\arguments{
\item{x}{numeric vector to reconcile.}
}
\value{
\code{x}
}
\description{
"Internal" function that \code{r2c} uses for book-keeping of what expressions need
to be reconciled across branches so that post-branch expressions can address
their side effects (i.e. assignments) or results irrespective of what branch
was taken at run time.  This function is exposed only so that users can
understand what it is should they see it in processed \code{r2c} calls.
}
