% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{reuse_calls}
\alias{reuse_calls}
\title{Identify Repeated Calls and Reuse First Instance}
\usage{
reuse_calls(x)
}
\arguments{
\item{x}{a call}
}
\value{
a list with elements:
\itemize{
\item x: the call with the re-used expressions substituted
\item reuse: named list with names the variables that reference the expressions
that are substituted, and values those expressions.
}
}
\description{
Complex statistics often re-use a simpler statistic multiple times, providing
the opportunity to optimize them by storing the result of the simple
statistic instead of recomputing it.  This function detects reuses of
sub-calls and modifies the call tree to implement the store-and-reuse
optimization.
}
\examples{
x <- runif(100)
y <- runif(100)
slope <- quote(((x - mean(x)) * (y - mean(y))) / (x - mean(x))^2)
(slope.r <- reuse_calls(slope))
identical(eval(slope), eval(slope.r))

intercept <- quote(
  mean(y) - mean(x) * ((x - mean(x)) * (y - mean(y))) / (x - mean(x))^2
)
(intercept.r <- reuse_calls(intercept))
identical(eval(intercept), eval(intercept.r))
}
