% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_code}
\alias{run_code}
\alias{run_code2}
\alias{run_code3}
\alias{run_group}
\title{Experimental Code Runner}
\usage{
run_code(code, fun, x)

run_code2(code, fun, x)

run_code3(code, fun, x, g)

run_group(shlib, fun, x, g)
}
\description{
Experimental Code Runner
}
