% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2c-package.R
\name{runners}
\alias{runners}
\title{Runner Functions}
\description{
"r2c_fun" functions generated by \code{\link{r2c}} can be invoked iteratively on
varying data with runner functions.  There are currently two types of
iteration available:
\itemize{
\item Group iteration with \code{\link{group_exec}}.
\item Rolling window iteration with \code{\link{rolli_exec}} and the other
\code{\link[=rollbw_exec]{roll*_exec}} functions.
}
}
\seealso{
Other r2c-topics: 
\code{\link{r2c-control-structures}},
\code{\link{r2c-expression-types}},
\code{\link{r2c-memory}},
\code{\link{r2c-package}},
\code{\link{r2c-performance}},
\code{\link{r2c-preprocess}},
\code{\link{r2c-supported-funs}}
}
\concept{r2c-topics}
\concept{runner functions}
\keyword{internal}
