% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-pow.R
\name{square}
\alias{square}
\title{Raise a Vector to the Power of Two}
\usage{
square(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
\code{x}, squared
}
\description{
Implemented as \code{x * x} to match what R does with \code{x ^ 2}.
}
