% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-subset.R
\name{subassign}
\alias{subassign}
\title{Assign to a Subset of a Vector}
\usage{
subassign(x, s, y)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
\code{x}, squared
}
\description{
"Internal" function that implement subassignment (e.g. \code{x[s] <- y}) for
\code{r2c}.  Standard R subassignments are converted to \code{subassign} so that the
\code{r2c} code processor can identify them as a sub-assignment instead of an
a subset nested in an assignment.
}
\details{
Unlike the R counterpart, or the \code{r2c} usage of this function,
regular R usage of this function does not modify \code{x} outside of the internal
scope of the function.  This function is an implementation detail and is
documented only for the rare cases where it becomes visible to the user.
}
