% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-unary.R
\name{uplus}
\alias{uplus}
\alias{uminus}
\title{Unary Operator Equivalent Functions}
\usage{
uplus(x)

uminus(x)
}
\arguments{
\item{x}{a logical numeric vector}
}
\value{
\code{x} for \code{uplus} (possibly coerced to integer/numeric), \code{-x} for
\code{uminus}
}
\description{
Required so that the compiler internals can easily disambiguate unary
plus/minus from the binary forms.  These should never be needed by the user,
but they are generated in the compilation step so that the C counterparts can
be used by \code{{r2c}}.
}
