% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc.R
\name{vec_eqlen}
\alias{vec_eqlen}
\title{Can All Vectors Be Considered Equal Size}
\usage{
vec_eqlen(sizes, groups, gmax, gmin)
}
\arguments{
\item{sizes}{vector of size values (see \code{alloc_dat})}

\item{groups}{vector of group designations (see \code{alloc_dat})}

\item{gmax}{scalar maximum iteration varying iteration size}

\item{gmin}{scalar minimum iteration varying iteration size}
}
\description{
Each element in sizes and groups represents one parameter.
}
