% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc.R
\name{alloc}
\alias{alloc}
\title{Allocate Required Storage}
\usage{
alloc(x, data, gmax, par.env, MoreArgs)
}
\arguments{
\item{x}{the result of preprocessing an expression}
}
\description{
Group data goes first, then result, then any other temporary or external
data.  For each call, we record indices into the storage list for each
argument and the result of evaluating the call.
}
