% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{get_c_code}
\alias{get_c_code}
\alias{get_r_code}
\alias{get_so_loc}
\alias{get_compile_out}
\alias{show_c_code}
\title{Extract Data from "r2c_fun" Objects}
\usage{
get_c_code(r2c.fun, all = TRUE)

get_r_code(r2c.fun)

get_so_loc(r2c.fun)

get_compile_out(r2c.fun)

show_c_code(r2c.fun, all = FALSE)
}
\arguments{
\item{r2c.fun}{an "r2c_fun" function as generated by e.g. \code{\link{r2c}}.}

\item{all}{TRUE or FALSE whether to retrieve all of the C code, or just
the portion directly corresponding to the primary runner function.}
}
\value{
for \code{get_r_code} a list with on or two members, the first "original"
is the R language object provided to the \link[=r2c-compile]{compilation functions}, the second "processed" is the version that the C
code is based on.  For all other functions a character vector, invisibly
for \code{show_c_code}.
}
\description{
"r2c_fun" functions contain embedded data used by the runners to call the
compiled native code associated with the functions.  The \verb{get_*} functions
documented here extract various aspects of this data.
}
\details{
\itemize{
\item \code{get_so_loc} the file system location of the shared object file.
\item \code{get_c_code} the generated C code used to produce the shared object.
\item \code{get_r_code} the R call that was translated into the C code; if
optimization was turned on it may have been modified from the original into
a semantically equivalent but in theory more efficient version (see
\code{\link{r2cq}}).
\item \code{get_compile_out} the "stdout" produced during the compilation of the
shared object.
\item \code{show_c_code} retrieves code with \code{get_c_code} and outputs to screen the
portion corresponding to the compiled expression, or optionally all of it.
}
}
\examples{
r2c_sum_sub <- r2cq(sum(x + y))
get_r_code(r2c_sum_sub)
show_c_code(r2c_sum_sub)
}
\seealso{
\code{\link{r2c-compile}}.
}
