% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{get_c_code}
\alias{get_c_code}
\alias{get_r_code}
\alias{get_so_loc}
\title{Extract Data from "r2c_fun" Objects}
\usage{
get_c_code(r2c.fun)

get_r_code(r2c.fun)

get_so_loc(r2c.fun)
}
\arguments{
\item{r2c.fun}{an "r2c_fun" function as generated by e.g. \code{\link{r2c}}.}
}
\value{
for \code{get_c_code} a character vector with each element a line in the
generated C code that correspond to the associated shared object, for
\code{get_r_code} the R language object the C code is based on, for \code{get_so_loc}
the file system location the shared object was stored in when generated by
\code{\link{r2c}}.
}
\description{
"r2c_fun" functions contain embedded data used by the runners to call the
compiled native code associated with the functions.  The \verb{get_*} functions
documented here extract various aspects of this data.
}
\examples{
r2c_sum_add <- r2cq(sum(x + y))
get_r_body(r2c_sum_add)
writeLines(get_c_code(r2c_sum_add))
}
\seealso{
\code{\link{r2c}}.
}
