% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc.R
\name{prepare}
\alias{prepare}
\title{Prepare Calls for Evaluation}
\usage{
prepare(call, data, g, env, depth, temp)
}
\arguments{
\item{call}{an unevaluated R call}

\item{data}{character the names of the data columns}

\item{g}{max group size}

\item{env}{calling frame to evaluate expression in}

\item{depth}{how deep into the expression we've recursed}

\item{temp}{structure that tracks the required size}
}
\value{
a list containing the accrued temporary storage requirements, as well
as a recursive list of entry point names along with the parameters to call
them with and the expected result size.
the entry point name, the parameters to call it
with (mixture of eval and uneval)
}
\description{
Retrieve appropriate entry points based on configuration parameters, and
compute required temporary storage sizes.
}
