% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc.R, R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Prepare Calls for Evaluation}
\usage{
preprocess(call, data, gmax)

preprocess(call, data, gmax)
}
\arguments{
\item{call}{an unevaluated R call}

\item{data}{character the names of the data columns}

\item{gmax}{max group size}

\item{env}{calling frame to evaluate expression in}

\item{depth}{how deep into the expression we've recursed}

\item{alloc.dat}{structure that tracks the required size}

\item{code}{a list of generated code data, one for each sub-call.}

\item{datai}{a list of integer vectors that represent indices into
\code{alloc.dat}, one for each sub-call.}

\item{ctrl}{a list of lists, each sub-list a set of R objects to pass on as
the control parameters, one for each sub-call.}
}
\value{
a list containing (update):
\itemize{
\item Accrued temporary storage requirements.
\item Size of the current call being assessed.
\item A linear list of generated code and associated calls.
}

a list containing (update):
\itemize{
\item Accrued temporary storage requirements.
\item Size of the current call being assessed.
\item A linear list of generated code and associated calls.
}
}
\description{
Retrieve appropriate entry points based on configuration parameters, and
compute required temporary storage sizes.

Retrieve appropriate entry points based on configuration parameters, and
compute required temporary storage sizes.
}
