% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2c-package.R
\docType{package}
\name{r2c-package}
\alias{r2c-package}
\alias{r2c}
\title{Fast Iterated Statistic Computation in R}
\description{
Compiles a subset of R into machine code so that expressions composed with
that subset can be applied repeatedly on varying data without interpreter
overhead.
}
\details{
For a quick start:
\itemize{
\item Review \link[=r2c-supported-funs]{supported functions}.
\item Use \link[=r2c-compile]{the compilation functions} to compile R.
\item Execute your \code{r2c} functions \link[=runners]{iteratively with runners}:
\itemize{
\item By \link[=group_exec]{group}.
\item Across \link[=rolli_exec]{windows}.
}
}

Details:
\itemize{
\item \link[=r2c-performance]{Performance considerations}.
\item \link[=r2c-memory]{Memory use}.
\item \link[=r2c-preprocess]{Preprocessing}.
\item Experimental support for \link[=r2c-control-structures]{control structures}.
\item Iteration \link[=r2c-expression-types]{constant and varying expressions}.
}
}
