% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2c-package.R
\docType{package}
\name{r2c-package}
\alias{r2c-package}
\title{Fast Iterated Statistic Computation in R}
\description{
Compiles a selected subset of R into native instructions so that that
expressions composed from that subset can be executed repeatedly on varying
data without interpreter overhead.
}
