% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2c-package.R
\docType{package}
\name{r2c-package}
\alias{r2c-package}
\title{Fast Iterated Statistic Computation in R}
\description{
Compiles a subset of R into machine code so that expressions composed with
that subset can be applied repeatedly on varying data without interpreter
overhead.  Use \code{\link{r2c}} to compile R, and \code{\link{group_exec}} or \code{\link{rolli_exec}} /
\code{\link{rollby_exec}} to run the result on groups or rolling windows of data.
}
