% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2c-package.R
\name{runners}
\alias{runners}
\title{Runner Functions}
\description{
"r2c_fun" functions generated by \code{\link{r2c}} can be invoked iteratively on
varying data with runner functions.  There are currently two types of
iteration available:
\itemize{
\item Group iteration with \code{\link{group_exec}}.
\item Rolling window iteration with \code{\link{rolli_exec}} and the other
\code{\link[=rollbw_exec]{roll*_exec}} functions.
}
}
\seealso{
\link[=r2c]{Package overview} for other \code{r2c} concepts.
}
\concept{runner functions}
\keyword{internal}
